refactor: Keep SqlHints and LockBehaviorBase in Internal; sharpen the #244 rule
Reverts the wave-3 moves of SqlHints and LockBehaviorBase. Both are
statement decoration, not query content: an optimizer hint does not
change the result set, and a lock behavior governs concurrency, not
what rows come back. Their naming flows were the weakest of the wave
(a shared hint constant can hold the string and re-wrap with Hints(s);
runtime lock selection has an additive post-1.0 escape — a new public
enum plus a ForUpdate overload), so the one-way-door cost of shipping
them at 1.0 is not justified.

Sharpens the boundary rule to three conjuncts (CLAUDE.md, ADR 0005):
a type leaves Internal only when it is a query's CONTENT (relation,
value, predicate, sort item, handle — not clause syntax, not statement
decoration, not a pending intermediate), a mainstream flow must write
its name, and no root type already names it. This is a cleaner test
than 'values vs clauses' and excludes hints/locks by the content
conjunct.

Final root-namespace set: SqlExpression, SqlCondition, ISubquery,
TableReference, SortOrder, ExpressionAlias, CommonTableExpression,
DbSequence. Drops the two SqlHints/LockBehaviorBase naming tests
(554 unit).

Refs #244
