+++ src/SqlArtisan/Internal/SqlPart/Clause/ForUpdate/LockBehaviorBase.cs
@@ -1,0 +1,5 @@
+namespace SqlArtisan.Internal;
+
+public abstract class LockBehaviorBase : SqlPart
+{
+}
