+++ src/SqlArtisan/Internal/SqlPart/Clause/SqlHints.cs
@@ -1,0 +1,14 @@
+namespace SqlArtisan.Internal;
+
+public sealed class SqlHints : SqlPart
+{
+    private readonly string _hints;
+
+    internal SqlHints(string hints)
+    {
+        _hints = hints;
+    }
+
+    internal override void Format(SqlBuildingBuffer buffer) =>
+        buffer.Append(_hints);
+}
